+++ moving_average.py
@@ -39,7 +39,7 @@ class MovingAverage:
     def init(self, initVal=None):
         if initVal is None:
             initVal = 0. 
-        self._ring_buffer = np.full(self._average_width, initVal, dtype=np.float32)        
+        self._ring_buffer = np.full(self._average_width, initVal, dtype=float)        
         self._average	= initVal;	
         self._sigma2	= 0
         self._is_init	= True;
